Trashy Master Tickets
This was a morning practice session to stay fresh branching and looping as well as raising exceptions. A 'Trashy Master Ticket' was created that lets the user know how many tickets are available, the total cost, and asks to proceed. It deliminates the number of tickets purchased and ends the program when the value is less than 1 and corrects for user input errors.